fix DeprecationWarning with datetime.utcnow() (#86)
This is not the same as
https://github.com/samuelcolvin/dirty-equals/issues/71, but it
originates at the same point. When I updated pytz to 2023.3.post1, I
started getting:
```
E       DeprecationWarning: datetime.datetime.utcnow() is deprecated and scheduled for removal in a future version. Use timezone-aware objects to represent datetimes in UTC: datetime.datetime.now(datetime.UTC).
```
This fixes it.

---------

Co-authored-by: Samuel Colvin <s@muelcolvin.com>